Add files via upload
A survey on the comparison of the Critical Path Method (CPM) with non-deterministic search algorithms to find a measure that indicates which algorithm is the best for a given scenario and to help future projects decide which method would work best for their case. So far the results of a heuristic analysis and some uninformed search algorithms has been acquired to be compared to the CPM on different levels in the next phase.